# Lift-Simple/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lift Simple API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List workouts for current user
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create a new workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutInput"
      responses:
        "201":
          description: Workout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get a workout with all exercises and sets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full workout detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutDetail"
        "404":
          description: Not found
    patch:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update a workout title or date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutUpdate"
      responses:
        "200":
          description: Updated workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        "404":
          description: Not found
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /workouts/{id}/complete:
    post:
      operationId: completeWorkout
      tags: [workouts]
      summary: Mark a workout as complete and get encouraging summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout summary with encouraging message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSummary"
        "404":
          description: Not found
  /workouts/{id}/exercises:
    post:
      operationId: addExerciseToWorkout
      tags: [workouts]
      summary: Add an exercise to a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutExerciseInput"
      responses:
        "201":
          description: Exercise added with last-time sets for reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutExerciseDetail"
        "404":
          description: Not found
  /workouts/{id}/exercises/{workoutExerciseId}:
    delete:
      operationId: removeExerciseFromWorkout
      tags: [workouts]
      summary: Remove an exercise from a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: workoutExerciseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found

  # Sets
...
components:
  schemas:
...
    Exercise:
      type: object
      required: [id, name, category, unit, isCustom, movementType]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [lower_body, upper_body, back, shoulders, arms, core, full_body]
        notes:
          type: ["string", "null"]
        unit:
          type: string
          enum: [lbs, kg]
        isCustom:
          type: boolean
        userId:
          type: ["integer", "null"]
        targetRepMin:
          type: integer
          default: 6
        targetRepMax:
          type: integer
          default: 12
        movementType:
          type: string
          enum: [upper_body, lower_body]
          default: upper_body
...
    Workout:
      type: object
      required: [id, userId, date, isCompleted, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        date:
          type: string
          format: date
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WorkoutInput:
      type: object
      required: [date]
      properties:
        title:
          type: string
        date:
          type: string
          format: date
    WorkoutUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        date:
          type: string
          format: date
    WorkoutSet:
      type: object
      required: [id, workoutExerciseId, weight, reps, setOrder]
      properties:
        id:
          type: integer
        workoutExerciseId:
          type: integer
        weight:
          type: number
          minimum: 0
        reps:
          type: integer
          minimum: 1
          maximum: 9999
        notes:
          type: ["string", "null"]
        setOrder:
          type: integer
        estimatedOneRepMax:
          type: ["number", "null"]
        isCompleted:
          type: boolean
          default: false
...
    WorkoutExerciseInput:
      type: object
      required: [exerciseId]
      properties:
        exerciseId:
          type: integer
    WorkoutExerciseDetail:
      type: object
      required: [id, workoutId, exerciseId, exercise, sets, lastTimeSets]
      properties:
        id:
          type: integer
        workoutId:
          type: integer
        exerciseId:
          type: integer
        exercise:
          $ref: "#/components/schemas/Exercise"
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
        lastTimeSets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
    WorkoutDetail:
      type: object
      required: [id, userId, date, isCompleted, createdAt, exercises]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        date:
          type: string
          format: date
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExerciseDetail"
    WorkoutSummary:
      type: object
      required: [workoutId, message, highlights]
      properties:
        workoutId:
          type: integer
        message:
          type: string
        highlights:
          type: array
          items:
            type: string
        totalSets:
          type: integer
        totalExercises:
          type: integer

    # Progress schemas
...