# Lift-Simple/lib/api-zod/src/generated/types/workoutExerciseInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lift Simple API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutExerciseInput {
  exerciseId: number;
}